Make `required()` on container args mean one-or-more.
This does two changes..

A `cardinality(X,0)` is now interpreted as being
X-or-more allowed arguments. I.e. it means that only a lower bound can
be set and interpreted as the max being unbounded.

Calling `required()` on a container bound argument makes it
accept 1-or-more arguments using a `cardinality(1,0)`.

Fixes #11