Don't assume tokenizer is correct about arg types.
When parsing opt and args don't assume the tokenizer
is correct in guessing the type or args. As it will
mislabel negative numbers as options since it doesn't
have the context from parsing. Instead opt and args
will try and interpret whatever tokens it can and only
fail the parse at that point if invalid values are present.

Fixes #14