Fix `-abc` opt names incosistency.
This fixes specifying short option names of `-abc` and not being able to
actually use them. The fix is to introduce a feature that mirrors the
compacting of short option when used to also behave the same when
specified. I.e. that `-abc` is split to multiple short options names.