Nicer help text formatting.
This brings back the 2 column layout of the help text options. It's
minimal in the formatting. And mirrors the clang style formatting
of wrapping the second column to the next line. And of leaving the
text description to wrap to the terminal.

Eventually will replace with something nicer that takes into account
LTR languages and such.

fixes #10