Fix command arguments ignored even when required.
When an argument of a subcommand is required it was being ignored
silently. The fix involed changing the group where command arguments
are added to and fixing the sequential arguments parser to check all
arguments instead of giving up on optional ones.

Semantic change for this fix is that arguments will signal matched
against no tokens (i.e. empty matching) if there are also no arguments
to match.

fixes #38