Fix op| loosing the group subtype.
The op| was assuming that the left value was *only* an arguments type.
Which meant that for subtypes of arguments it would copy-slice the
subtype away. This would cause group definitions to get lost. With that
fixed it was now possible to fix the downstream issue of sequential args
causing infinite parsing loops when all the args where optional. Now the
all optional args case correctly causes an unmatched parse instead of
the infinite loop.

fixes #73