Add help print customization.
Adds ability to customize help printing customization. Makes it possible
to avoid iostreams for printing help.

Also removes use of algorithm. The use of algos was minor and the
equivalent code turns out to be simpler and more efficient. This avoids
the inclusion of the heavy weight header as it has become with
recent C++ standards.

fixes #50