Use more specific detection of optional.
This changes detection of `std::optional` to use the feature macro *after* including the optional header. This, hopefully, allows for implementations that do not set the general __cplusplus macro to match the implementation. This also adds a message to the corresponding test to indicate when optional support is not detected.